refactor(payments): merge assigned-orders into received-orders [GH-000]
The last four cross-feature imports were payments: assigned-orders and
received-orders importing each other. Counting the files settled what the
import graph only hinted at.

received-orders held fourteen files: the card, its three sibling components,
the actions hook, the types, the constants, the queries. assigned-orders held
six, and not one of them was a component, a type, or a domain rule -- it was a
query call and a page. Even its own fetchAssignedOrders already lived in
received-orders' query file.

So there were never two features. Assigned orders are received orders under a
delegation filter, and the second directory was a view wearing a feature's
clothes. Merging them removes the import cycle by removing the thing that made
it a cycle, rather than by relocating imports until the check stops counting.

received-orders now exports two pages. Test ids and the ASSIGNED_ORDERS_QUERY_KEY
string keep the word "assigned" -- those name a view and a cache entry, not a
directory, and renaming them would churn e2e specs for nothing.

The boundaries ratchet drops from 4 to 0, so any new cross-feature import now
fails the build outright.

Verified: typecheck, lint (0/0), payments 554 tests, format:check, all four
custom gates, and a full build with /[locale]/assigned confirmed present in the
route table.
