--- apps/payments/src/features/assigned-orders/index.ts
@@ -1,1 +1,0 @@
-export { AssignedOrdersPage } from "./presentation/pages/AssignedOrdersPage";
