+++ apps/payments/src/features/received-orders/domain/searchParams.ts
@@ -1,9 +1,20 @@
 import { parseAsStringEnum } from "nuqs";
 
-import { FILTER_STATUSES, type FilterStatus } from "./constants";
+import {
+  ASSIGNED_FILTER_STATUSES,
+  FILTER_STATUSES,
+  type AssignedFilterStatus,
+  type FilterStatus,
+} from "./constants";
 
 export const receivedOrdersSearchParams = {
   filter: parseAsStringEnum<FilterStatus>([...FILTER_STATUSES]).withDefault(
     "all",
   ),
+};
+
+export const assignedOrdersSearchParams = {
+  filter: parseAsStringEnum<AssignedFilterStatus>([
+    ...ASSIGNED_FILTER_STATUSES,
+  ]).withDefault("actionable"),
 };
